+++ shiny/express/_output.py
@@ -8,13 +8,8 @@ from typing import Callable, Generator, TypeVar, overload
 from .. import ui
 from .._typing_extensions import ParamSpec
 from ..render.renderer import RendererBase, RendererBaseT
-from ..render.transformer import OutputRenderer
-from ..render.transformer._transformer import OT
 
-__all__ = (
-    "ui_kwargs",
-    "suspend_display",
-)
+__all__ = ("suspend_display",)
 
 P = ParamSpec("P")
 R = TypeVar("R")
@@ -22,7 +17,7 @@ CallableT = TypeVar("CallableT", bound=Callable[..., object])
 
 
 # TODO-barret-future; quartodoc entry?
-def ui_kwargs(
+def output_args(
     **kwargs: object,
 ) -> Callable[[RendererBaseT], RendererBaseT]:
     """
@@ -31,9 +26,9 @@ def ui_kwargs(
     Each Shiny render function (like :func:`~shiny.render.plot`) can display itself when
     declared within a Shiny inline-style application. In the case of
     :func:`~shiny.render.plot`, the :func:`~shiny.ui.output_plot` function is called
-    implicitly to display the plot. Use the `@ui_kwargs` decorator to specify
-    arguments to be passed to `output_plot` (or whatever the corresponding UI function
-    is) when the render function displays itself.
+    implicitly to display the plot. Use the `@ui_kwargs` decorator to specify arguments
+    to be passed to `output_plot` (or whatever the corresponding UI function is) when
+    the render function displays itself.
 
     Parameters
     ----------
@@ -47,51 +42,7 @@ def ui_kwargs(
     """
 
     def wrapper(renderer: RendererBaseT) -> RendererBaseT:
-        # renderer._default_ui_args = args
-        renderer._default_ui_kwargs = kwargs
-        return renderer
-
-    return wrapper
-
-
-def output_args(
-    *args: object,
-    **kwargs: object,
-) -> Callable[[OutputRenderer[OT]], OutputRenderer[OT]]:
-    """
-    Sets default UI arguments for a Shiny rendering function.
-
-    Each Shiny render function (like :func:`~shiny.render.plot`) can display itself when
-    declared within a Shiny inline-style application. In the case of
-    :func:`~shiny.render.plot`, the :func:`~shiny.ui.output_plot` function is called
-    implicitly to display the plot. Use the `@output_args` decorator to specify
-    arguments to be passed to `output_plot` (or whatever the corresponding UI function
-    is) when the render function displays itself.
-
-
-    Parameters
-    ----------
-    *args
-        Positional arguments to be passed to the UI function.
-    **kwargs
-        Keyword arguments to be passed to the UI function.
-
-    Returns
-    -------
-    :
-        A decorator that sets the default UI arguments for a Shiny rendering function.
-    """
-
-    def wrapper(renderer: OutputRenderer[OT]) -> OutputRenderer[OT]:
-        if not isinstance(renderer, OutputRenderer):
-            raise TypeError(
-                f"Expected an OutputRenderer, but got {type(renderer).__name__}."
-                "\nIf you are trying to set default UI arguments for a `Renderer`, use"
-                " `@ui_kwargs` instead."
-            )
-        renderer._default_ui_args = args
-        renderer._default_ui_kwargs = kwargs
-
+        renderer._auto_output_ui_kwargs = kwargs
         return renderer
 
     return wrapper
@@ -152,7 +103,7 @@ def suspend_display(
     # display yourself"
     if isinstance(fn, RendererBase):
         # By setting the class value, the `self` arg will be auto added.
-        fn.default_ui = null_ui
+        fn.auto_output_ui = null_ui
         return fn
 
     return suspend_display_ctxmgr()(fn)