+++ shiny/templates/package-templates/js-output/custom_component/custom_component.py
@@ -27,7 +27,7 @@ class render_custom_component(Renderer[int]):
     """
 
     # The UI used within Shiny Express mode
-    def default_ui(self, id: str) -> Tag:
+    def auto_output_ui(self, id: str) -> Tag:
         return custom_component(id, height=self.height)
 
     # The init method is used to set up the renderer's parameters.